Introduced size primitve types (usize and isize) Language: Added `usize` and `isize` primitive types (with `us` and `is` suffixes respectively) with size equal to `sizeof(void*)` (architecture dependent). Language: Array access and array/string size operations now use `usize` instead of `u64`.

+++ core/include/linc/bound_tree/BoundAccessExpression.hpp
@@ -8,7 +8,7 @@ namespace linc
     class BoundAccessExpression final : public BoundExpression
     {
     public:
-        BoundAccessExpression(std::unique_ptr<const BoundExpression> base, Types::u64 index, const Types::type& type)
+        BoundAccessExpression(std::unique_ptr<const BoundExpression> base, Types::usize index, const Types::type& type)
             :BoundExpression(type), m_base(std::move(base)), m_index(index)
         {}
 
@@ -18,13 +18,13 @@ namespace linc
         }
 
         [[nodiscard]] inline const BoundExpression* const getBase() const { return m_base.get(); }
-        [[nodiscard]] inline Types::u64 getIndex() const { return m_index; }
+        [[nodiscard]] inline Types::usize getIndex() const { return m_index; }
     private:
         virtual std::string toStringInner() const final override
         {
             return "Access Expression";
         }
         const std::unique_ptr<const BoundExpression> m_base; 
-        const Types::u64 m_index;
+        const Types::usize m_index;
     };
 }
